tint: Lex different tokens for float suffixes
Generate different tokens for:
• 'f' suffixed float literals
• no-suffix integer literals

'f' and no-suffix are currently both treated as f32 by the resolver,
but this is the first step to supporting abstract floats.

Bug: tint:1504

+++ src/tint/resolver/resolver_constants.cc
@@ -38,12 +38,12 @@ sem::Constant Resolver::EvaluateConstantValue(const ast::LiteralExpression* lite
         literal,
         [&](const ast::IntLiteralExpression* lit) {
             if (lit->suffix == ast::IntLiteralExpression::Suffix::kU) {
-                return sem::Constant{type, {u32(static_cast<uint32_t>(lit->value))}};
+                return sem::Constant{type, {u32(lit->value)}};
             }
-            return sem::Constant{type, {i32(static_cast<int32_t>(lit->value))}};
+            return sem::Constant{type, {i32(lit->value)}};
         },
         [&](const ast::FloatLiteralExpression* lit) {
-            return sem::Constant{type, {lit->value}};
+            return sem::Constant{type, {f32(lit->value)}};
         },
         [&](const ast::BoolLiteralExpression* lit) {
             return sem::Constant{type, {lit->value}};
